refactor-02: Data Models Consolidation + using conftest for Python path setup
