+++ tests/core/test_chat_with_prometheus.py
@@ -6,8 +6,7 @@ from unittest.mock import Mock, patch
 import sys
 import os
 
-# Add src to path for imports
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..', 'src'))
+# Python path is now handled by conftest.py
 
 
 class TestSemanticAnalysis:
