+++ tests/mcp_server/test_mcp.py
@@ -3,8 +3,7 @@
 import sys
 import os
 
-# Add src to path for imports
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..', 'src'))
+# Python path is now handled by conftest.py
 
 
 def test_observability_mcp_server_imports():
