+++ tests/mcp_server/test_prometheus_tools.py
@@ -5,8 +5,7 @@ import json
 import sys
 import os
 
-# Add src to path for imports
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..', 'src'))
+# Python path is now handled by conftest.py
 
 
 class TestPrometheusToolsBasic:
